+++ CHANGELOG.md
@@ -2,13 +2,14 @@
 
 ## 0.1.4
 - Sorting contexts now respects regions
-- Support for sorting methods added.
+- Support for sorting methods added
+- Support for sorting interfaces added
 
 ## 0.1.3
 - Sort Properties now organizes by modifier first in the following order: public, internal, protected internal, protected, private
 
 ## 0.1.2
-- Sort action menu options no longer appear when there is one or fewer properties or initializer members. 
+- Sort action menu options no longer appear when there is one or fewer properties or initializer members.
 
 ## 0.1.1
 - Fixed minor issue with Object Initializer sort showing option for Class property sort as well.